refactor: rename graph_highlight to graphHighlight
