+++ app/javascript/javascript/graphHighlight.js
